#同步RuoYi更新-修复导入Excel时字典字段类型为Long转义为空问题

+++ oa-common/oa-common-core/src/main/java/com/oa/common/core/utils/poi/ExcelUtil.java
@@ -294,7 +294,7 @@ public class ExcelUtil<T>
                             }
                         }
                     }
-                    else if ((Integer.TYPE == fieldType || Integer.class == fieldType) && StringUtils.isNumeric(Convert.toStr(val)))
+                    else if ((Long.TYPE == fieldType || Long.class == fieldType) && StringUtils.isNumeric(Convert.toStr(val)))
                     {
                         val = Convert.toInt(val);
                     }
